Add support for Scientific Linux in Yum repo
This commit uses the $::operatingsystemmajrelease fact for managing the
Yum repository on Scientific Linux.  Yum's "$releasever" resolves to
"6.6" (major.minor) on Scientific Linux, where it resolves to "major" or
"major(.server)?" on RedHat.  That causes the repo to fail on Scientific
Linux.

The other solution would be to add symlinks for major.minor in the
upstream Yum repo.  E.g. 6.6 => 6