Add nginx example
Without this, there is not an example showing how nginx can be used with
this module.  The example here is what it took to get the icingaweb2 app
running behind nginx using the voxpupuli-nginx module in combination
with this module, and some php-fpm configuration.  Perhaps this can help
someone else.