Add parameter AddCGIPassAuth to Apache config
If PHP FPM is used for Icingaweb2, mod_proxy_fcgi on Debian will, by
default, strip Authentication Headers from the request.
While this does not influence the "normal" interaction with the website,
it will prohibit using the API functionality in an automated fashion (,
where Basic Authentication is used instead of cookies).